Addressing review comments.
