+++ .github/workflows/fuzzing-build.yaml
@@ -34,9 +34,6 @@ jobs:
             image: connectedhomeip/chip-build:0.5.48
             volumes:
                 - "/tmp/log_output:/tmp/test_logs"
-            options:
-                --sysctl "net.ipv6.conf.all.disable_ipv6=0
-                net.ipv4.conf.all.forwarding=1 net.ipv6.conf.all.forwarding=1"
 
         steps:
             - name: Checkout
