+++ examples/all-clusters-app/linux/BUILD.gn
@@ -14,17 +14,18 @@
 
 import("//build_overrides/build.gni")
 import("//build_overrides/chip.gni")
+import("${build_root}/config/compiler/compiler.gni")
 
 import("${chip_root}/src/lib/lib.gni")
 
-executable("chip-all-clusters-app") {
+source_set("chip-all-clusters-common") {
   sources = [
     "${chip_root}/examples/all-clusters-app/all-clusters-common/src/binding-handler.cpp",
     "${chip_root}/examples/all-clusters-app/all-clusters-common/src/bridged-actions-stub.cpp",
     "${chip_root}/examples/all-clusters-app/all-clusters-common/src/static-supported-modes-manager.cpp",
     "include/tv-callbacks.cpp",
     "include/tv-callbacks.h",
-    "main.cpp",
+    "main-common.cpp",
   ]
 
   deps = [
@@ -41,10 +42,41 @@ executable("chip-all-clusters-app") {
   if (chip_build_libshell) {
     defines = [ "ENABLE_CHIP_SHELL" ]
   }
+}
 
-  output_dir = root_out_dir
+if (is_libfuzzer) {
+  executable("chip-all-clusters-app-fuzzing") {
+    sources = [ "fuzzing-main.cpp" ]
+
+    deps = [
+      ":chip-all-clusters-common",
+      "${chip_root}/examples/platform/linux:app-main",
+    ]
+
+    cflags = [ "-Wconversion" ]
+
+    output_dir = root_out_dir
+  }
+} else {
+  executable("chip-all-clusters-app") {
+    sources = [ "main.cpp" ]
+
+    deps = [
+      ":chip-all-clusters-common",
+      "${chip_root}/examples/platform/linux:app-main",
+    ]
+
+    cflags = [ "-Wconversion" ]
+
+    output_dir = root_out_dir
+  }
 }
 
 group("linux") {
-  deps = [ ":chip-all-clusters-app" ]
+  deps = []
+  if (is_libfuzzer) {
+    deps += [ ":chip-all-clusters-app-fuzzing" ]
+  } else {
+    deps += [ ":chip-all-clusters-app" ]
+  }
 }
